Updated to not re-run already analyzed rpms

+++ fedora.py
@@ -43,11 +43,11 @@ for repo in ctl.repo_list:
                 ctl.prep_rpm(repo, filename)
                 metadata = ctl.get_metadata(filename)
                 elfs = ctl.find_elfs()
-                if elfs:
+                results_file = results_dir + '/' + filename + '.json'
+                if elfs and not os.path.isfile(results_file):
                     results = ctl.scan_elfs(filename, elfs)
                     results_dir = output_dir + '/' + filename[0]
                     ctl.redteam.funcs.mkdir_p(results_dir)
-                    results_file = results_dir + '/' + filename + '.json'
                     with open(results_file, 'w') as f:
                         json.dump({'metadata': metadata,
                                    'results': results}, f, indent=4)
